backend/harvesters/Mastodon/requirements.txt
elasticsearch8==8.11.0
Mastodon.py==1.8.1